Change types in Filter classes (#1220)

+++ gitlab4j-models/src/main/java/org/gitlab4j/api/models/GroupFilter.java
@@ -15,7 +15,7 @@ import org.gitlab4j.models.utils.JacksonJson;
 public class GroupFilter implements Serializable {
     private static final long serialVersionUID = 1L;
 
-    private List<Integer> skipGroups;
+    private List<Long> skipGroups;
     private Boolean allAvailable;
     private String search;
     private GroupOrderBy orderBy;
@@ -33,7 +33,7 @@ public class GroupFilter implements Serializable {
      * @param skipGroups List of group IDs to not include in the search
      * @return the reference to this GroupFilter instance
      */
-    public GroupFilter withSkipGroups(List<Integer> skipGroups) {
+    public GroupFilter withSkipGroups(List<Long> skipGroups) {
         this.skipGroups = skipGroups;
         return (this);
     }
